+++ source/common/formatter/BUILD
@@ -54,7 +54,7 @@ envoy_cc_library(
         "//envoy/api:api_interface",
         "//envoy/http:protocol_interface",
         "//source/common/api:os_sys_calls_lib",
-        "//source/common/http:header_utility_lib",
+        "//source/common/http:header_validation_lib",
         "//source/common/http:utility_lib",
         "//source/common/protobuf:utility_lib",
         "//source/common/runtime:runtime_features_lib",
