+++ source/common/http/http1/BUILD
@@ -57,6 +57,7 @@ envoy_cc_library(
         "//source/common/http:exception_lib",
         "//source/common/http:header_map_lib",
         "//source/common/http:header_utility_lib",
+        "//source/common/http:header_validation_lib",
         "//source/common/http:headers_lib",
         "//source/common/http:status_lib",
         "//source/common/http:utility_lib",
