+++ source/common/http/http2/BUILD
@@ -53,6 +53,7 @@ envoy_cc_library(
         "//source/common/http:exception_lib",
         "//source/common/http:header_map_lib",
         "//source/common/http:header_utility_lib",
+        "//source/common/http:header_validation_lib",
         "//source/common/http:headers_lib",
         "//source/common/http:status_lib",
         "//source/common/http:utility_lib",
