+++ source/extensions/filters/common/mutation_rules/BUILD
@@ -17,6 +17,7 @@ envoy_cc_library(
         "//source/common/common:macros",
         "//source/common/common:regex_lib",
         "//source/common/http:header_utility_lib",
+        "//source/common/http:header_validation_lib",
         "//source/common/http:headers_lib",
         "@com_google_absl//absl/container:flat_hash_set",
         "@com_google_absl//absl/strings",
