+++ test/common/http/utility_test.cc
@@ -10,6 +10,7 @@
 #include "source/common/http/exception.h"
 #include "source/common/http/header_map_impl.h"
 #include "source/common/http/http1/settings.h"
+#include "source/common/http/protocol_options.h"
 #include "source/common/http/utility.h"
 #include "source/common/network/address_impl.h"
 
@@ -547,7 +548,7 @@ namespace {
 envoy::config::core::v3::Http2ProtocolOptions parseHttp2OptionsFromV3Yaml(const std::string& yaml) {
   envoy::config::core::v3::Http2ProtocolOptions http2_options;
   TestUtility::loadFromYamlAndValidate(yaml, http2_options);
-  return ::Envoy::Http2::Utility::initializeAndValidateOptions(http2_options).value();
+  return ::Envoy::Http2::ProtocolOptions::initializeAndValidateOptions(http2_options).value();
 }
 
 } // namespace
@@ -588,68 +589,6 @@ initial_connection_window_size: 65535
     EXPECT_EQ(65535U, http2_options.initial_stream_window_size().value());
     EXPECT_EQ(65535U, http2_options.initial_connection_window_size().value());
   }
-}
-
-TEST(HttpUtility, ValidateStreamErrors) {
-  // Both false, the result should be false.
-  envoy::config::core::v3::Http2ProtocolOptions http2_options;
-  EXPECT_FALSE(Envoy::Http2::Utility::initializeAndValidateOptions(http2_options)
-                   .value()
-                   .override_stream_error_on_invalid_http_message()
-                   .value());
-
-  // If the new value is not present, the legacy value is respected.
-  http2_options.set_stream_error_on_invalid_http_messaging(true);
-  EXPECT_TRUE(Envoy::Http2::Utility::initializeAndValidateOptions(http2_options)
-                  .value()
-                  .override_stream_error_on_invalid_http_message()
-                  .value());
-
-  // If the new value is present, it is used.
-  http2_options.mutable_override_stream_error_on_invalid_http_message()->set_value(true);
-  http2_options.set_stream_error_on_invalid_http_messaging(false);
-  EXPECT_TRUE(Envoy::Http2::Utility::initializeAndValidateOptions(http2_options)
-                  .value()
-                  .override_stream_error_on_invalid_http_message()
-                  .value());
-
-  // Invert values - the new value should still be used.
-  http2_options.mutable_override_stream_error_on_invalid_http_message()->set_value(false);
-  http2_options.set_stream_error_on_invalid_http_messaging(true);
-  EXPECT_FALSE(Envoy::Http2::Utility::initializeAndValidateOptions(http2_options)
-                   .value()
-                   .override_stream_error_on_invalid_http_message()
-                   .value());
-}
-
-TEST(HttpUtility, ValidateStreamErrorsWithHcm) {
-  envoy::config::core::v3::Http2ProtocolOptions http2_options;
-  http2_options.set_stream_error_on_invalid_http_messaging(true);
-  EXPECT_TRUE(Envoy::Http2::Utility::initializeAndValidateOptions(http2_options)
-                  .value()
-                  .override_stream_error_on_invalid_http_message()
-                  .value());
-
-  // If the HCM value is present it will take precedence over the old value.
-  ProtobufWkt::BoolValue hcm_value;
-  hcm_value.set_value(false);
-  EXPECT_FALSE(Envoy::Http2::Utility::initializeAndValidateOptions(http2_options, true, hcm_value)
-                   .value()
-                   .override_stream_error_on_invalid_http_message()
-                   .value());
-  // The HCM value will be ignored if initializeAndValidateOptions is told it is not present.
-  EXPECT_TRUE(Envoy::Http2::Utility::initializeAndValidateOptions(http2_options, false, hcm_value)
-                  .value()
-                  .override_stream_error_on_invalid_http_message()
-                  .value());
-
-  // The override_stream_error_on_invalid_http_message takes precedence over the
-  // global one.
-  http2_options.mutable_override_stream_error_on_invalid_http_message()->set_value(true);
-  EXPECT_TRUE(Envoy::Http2::Utility::initializeAndValidateOptions(http2_options, true, hcm_value)
-                  .value()
-                  .override_stream_error_on_invalid_http_message()
-                  .value());
 }
 
 TEST(HttpUtility, ValidateStreamErrorConfigurationForHttp1) {
