+++ test/integration/quic_http_integration_test.cc
@@ -1109,7 +1109,7 @@ TEST_P(QuicHttpIntegrationTest, ResetRequestWithInvalidCharacter) {
   codec_client_ = makeHttpConnection(lookupPort("http"));
 
   std::string value = std::string(1, 2);
-  EXPECT_FALSE(Http::HeaderUtility::headerValueIsValid(value));
+  EXPECT_FALSE(Http::HeaderValidation::headerValueIsValid(value));
   default_request_headers_.addCopy("illegal_header", value);
   auto encoder_decoder = codec_client_->startRequest(default_request_headers_);
   request_encoder_ = &encoder_decoder.first;
