+++ test/mocks/upstream/BUILD
@@ -18,7 +18,7 @@ envoy_cc_mock(
         "//envoy/upstream:upstream_interface",
         "//source/common/common:thread_lib",
         "//source/common/config:metadata_lib",
-        "//source/common/http:utility_lib",
+        "//source/common/http:protocol_options_lib",
         "//source/common/http/http1:codec_stats_lib",
         "//source/common/http/http2:codec_stats_lib",
         "//source/common/network:raw_buffer_socket_lib",
